+++ CHANGELOG.md
@@ -29,6 +29,7 @@ the following form: YYYY.0M.0D.
   project
 - Added information about how to make a pull request to every part of the
   project
+  - Added components to 01 homework
 
 ### Changed